refactor(server): extract document routes
